put secrets stuff in own module
